backend/requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
psycopg2-binary==2.9.9
pydantic==2.5.3
Faker==20.1.0

